requirements.txt
# Core
pandas>=2.0.0
openpyxl>=3.1.0
psycopg2-binary>=2.9.0

# Google Gemini
google-generativeai>=0.3.0

# File processing
python-pptx>=0.6.21
python-docx>=1.0.0
reportlab>=4.0.0
Pillow>=10.0.0
opencv-python>=4.8.0
PyMuPDF==1.23.8

# Pipeline
tqdm>=4.66.0
tenacity>=8.2.0
